pyproject.toml
[tool.poetry]
name = "vectorfriends-connect"
version = "0.1.0"
description = ""
authors = ["jalakoo <jalakoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
functions-framework = "^3.7.0"
basicauth = "^1.0.0"
neo4j-rust-ext = "^5.20.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
